Added some console.log statements to show how session works
before session middleware, req.session doesn't exist. Because the
client sends a cookie with a session id, when the session middleware
runs, it creates the req.sessions and restores whatever it was last.
This is how the express server knows how many times you have viewed
the page. When the server is restarted, that information is no longer
present, but the session ID cookie is still stored in the client's
browser.